requirements.txt
gradio
ipykernel
ipywidgets
openai>=1.0.0
motor
neo4j
pinecone
pymongo
python-dotenv
tenacity
tqdm
curl_cffi
beautifulsoup4
tiktoken
google-generativeai